+++ server/src/main/java/net/guides/springboot2/crud/controller/COController.java
@@ -23,57 +23,59 @@ import org.springframework.data.domain.Sort;
 @RestController
 @RequestMapping("/co")
 public class COController {
-    @Autowired
-    private CORepository coRepository;
+	@Autowired
+	private CORepository coRepository;
 
-    @GetMapping
-    public List<CO> getAllCOs() {
-        return coRepository.findAll(Sort.by(Sort.Direction.ASC, "id"));
-    }
+	@GetMapping
+	public List<CO> getAllCOs() {
+		return coRepository.findAll(Sort.by(Sort.Direction.ASC, "id"));
+	}
 
-    @GetMapping("{id}")
-    public ResponseEntity<CO> getCOById(@PathVariable(value = "id") Long coId) throws ResourceNotFoundException {
-        CO co = coRepository.findById(coId)
-                .orElseThrow(() -> new ResourceNotFoundException("CO not found for this id :: " + coId));
-        return ResponseEntity.ok().body(co);
-    }
+	@GetMapping("{id}")
+	public ResponseEntity<CO> getCOById(@PathVariable(value = "id") int coId) throws ResourceNotFoundException {
+		CO co = coRepository.findById(coId)
+				.orElseThrow(() -> new ResourceNotFoundException("CO not found for this id :: " + coId));
+		return ResponseEntity.ok().body(co);
+	}
 
-    @GetMapping("idc={id}")
-    public ResponseEntity<List<CO>> getCOByIdcontract(@PathVariable(value = "id") Long idcontract) throws ResourceNotFoundException {
-        List<CO> co = coRepository.findByIdcontract(idcontract);
-        return ResponseEntity.ok().body(co);
-    }
+	@GetMapping("idc={id}")
+	public ResponseEntity<List<CO>> getCOByIdcontract(@PathVariable(value = "id") int idcontract)
+			throws ResourceNotFoundException {
+		List<CO> co = coRepository.findByIdcontract(idcontract);
+		return ResponseEntity.ok().body(co);
+	}
 
-    @GetMapping("fp&idc={id}")
-    public ResponseEntity<List<CO>> getCOByIdcontractWhereNeplatit(@PathVariable(value = "id") Long idcontract) throws ResourceNotFoundException {
-        List<CO> co = coRepository.findByIdcontractAndTip(idcontract, "Concediu fără plată");
-        return ResponseEntity.ok().body(co);
-    }
+	@GetMapping("fp&idc={id}")
+	public ResponseEntity<List<CO>> getCOByIdcontractWhereNeplatit(@PathVariable(value = "id") int idcontract)
+			throws ResourceNotFoundException {
+		List<CO> co = coRepository.findByIdcontractAndTip(idcontract, "Concediu fără plată");
+		return ResponseEntity.ok().body(co);
+	}
 
-    @PostMapping
-    public CO createCO(@RequestBody CO co) {
-        return coRepository.save(co);
-    }
+	@PostMapping
+	public CO createCO(@RequestBody CO co) {
+		return coRepository.save(co);
+	}
 
-    @PutMapping("{id}")
-    public ResponseEntity<CO> updateCO(@PathVariable(value = "id") Long coId, @RequestBody CO coDetails)
-            throws ResourceNotFoundException {
-        CO co = coRepository.findById(coId)
-                .orElseThrow(() -> new ResourceNotFoundException("CO not found for this id :: " + coId));
+	@PutMapping("{id}")
+	public ResponseEntity<CO> updateCO(@PathVariable(value = "id") int coId, @RequestBody CO coDetails)
+			throws ResourceNotFoundException {
+		CO co = coRepository.findById(coId)
+				.orElseThrow(() -> new ResourceNotFoundException("CO not found for this id :: " + coId));
 
-        coDetails.setId(co.getId());
-        final CO updatedCO = coRepository.save(coDetails);
-        return ResponseEntity.ok(updatedCO);
-    }
+		coDetails.setId(co.getId());
+		final CO updatedCO = coRepository.save(coDetails);
+		return ResponseEntity.ok(updatedCO);
+	}
 
-    @DeleteMapping("{id}")
-    public Map<String, Boolean> deleteCO(@PathVariable(value = "id") Long coId) throws ResourceNotFoundException {
-        CO co = coRepository.findById(coId)
-                .orElseThrow(() -> new ResourceNotFoundException("CO not found for this id :: " + coId));
+	@DeleteMapping("{id}")
+	public Map<String, Boolean> deleteCO(@PathVariable(value = "id") int coId) throws ResourceNotFoundException {
+		CO co = coRepository.findById(coId)
+				.orElseThrow(() -> new ResourceNotFoundException("CO not found for this id :: " + coId));
 
-        coRepository.delete(co);
-        Map<String, Boolean> response = new HashMap<>();
-        response.put("deleted", Boolean.TRUE);
-        return response;
-    }
+		coRepository.delete(co);
+		Map<String, Boolean> response = new HashMap<>();
+		response.put("deleted", Boolean.TRUE);
+		return response;
+	}
 }
